refactor(plot): separate Bokeh class functions into classes (#279)
* chore(plot): rename message_flow_source to message_flow


* refactor(plot): delegate message_flow()in Bokeh to BokehMessageFlow class


* chore(plot): rename timeseries_source to timeseries


* refactor(plot): delegate timeseries()in Bokeh to BokehTimeSeries class


* chore(plot): rename stacked_bar_source to stacked_bar


* refactor(plot): delegate stacked_bar()in Bokeh to BokehStackedBar class


* chore(plot): rename callback_scheduling_source to callback_scheduling


* refactor(plot): delegate callback_scheduling()in Bokeh to BokehCallbackSched class


---------
