+++ src/caret_analyze/plot/visualize_lib/bokeh/timeseries.py
@@ -18,14 +18,82 @@ from __future__ import annotations
 from typing import Any, Dict, List, Optional, Sequence, Tuple, Union
 
 from bokeh.models import HoverTool
-from bokeh.plotting import ColumnDataSource
+from bokeh.plotting import ColumnDataSource, Figure
 
-from .util import HoverKeysFactory, HoverSource, LegendManager
-
-from ....record import RecordsInterface
+from .util import (apply_x_axis_offset, ColorSelectorFactory,
+                   HoverKeysFactory, HoverSource, init_figure, LegendManager)
+from ...metrics_base import MetricsBase
+from ....record import Range, RecordsInterface
 from ....runtime import CallbackBase, Communication, Publisher, Subscription
 
 TimeSeriesTypes = Union[CallbackBase, Communication, Union[Publisher, Subscription]]
+
+
+class BokehTimeSeries:
+
+    def __init__(
+        self,
+        metrics: MetricsBase,
+        xaxis_type: str,
+        ywheel_zoom: bool,
+        full_legends: bool
+    ) -> None:
+        self._metrics = metrics
+        self._xaxis_type = xaxis_type
+        self._ywheel_zoom = ywheel_zoom
+        self._full_legends = full_legends
+
+    def create_figure(self) -> Figure:
+        target_objects = self._metrics.target_objects
+        timeseries_records_list = self._metrics.to_timeseries_records_list(self._xaxis_type)
+
+        # Initialize figure
+        y_axis_label = timeseries_records_list[0].columns[1]
+        if y_axis_label == 'frequency':
+            y_axis_label = y_axis_label + ' [Hz]'
+        elif y_axis_label in ['period', 'latency']:
+            y_axis_label = y_axis_label + ' [ms]'
+        else:
+            raise NotImplementedError()
+        if isinstance(target_objects[0], CallbackBase):
+            title = f'Time-line of callbacks {y_axis_label}'
+        elif isinstance(target_objects[0], Communication):
+            title = f'Time-line of communications {y_axis_label}'
+        else:
+            title = f'Time-line of publishes/subscribes {y_axis_label}'
+        fig = init_figure(title, self._ywheel_zoom, self._xaxis_type, y_axis_label)
+
+        # Apply xaxis offset
+        records_range = Range([to.to_records() for to in target_objects])
+        frame_min, frame_max = records_range.get_range()
+        if self._xaxis_type == 'system_time':
+            apply_x_axis_offset(fig, frame_min, frame_max)
+
+        # Draw lines
+        color_selector = ColorSelectorFactory.create_instance(coloring_rule='unique')
+        legend_manager = LegendManager()
+        line_source = LineSource(legend_manager, target_objects[0], frame_min, self._xaxis_type)
+        fig.add_tools(line_source.create_hover())
+        for to, timeseries in zip(target_objects, timeseries_records_list):
+            renderer = fig.line(
+                'x', 'y',
+                source=line_source.generate(to, timeseries),
+                color=color_selector.get_color()
+            )
+            legend_manager.add_legend(to, renderer)
+
+        # Draw legends
+        num_legend_threshold = 20
+        """
+        In Autoware, the number of callbacks in a node is less than 20.
+        Here, num_legend_threshold is set to 20 as the maximum value.
+        """
+        legends = legend_manager.create_legends(num_legend_threshold, self._full_legends)
+        for legend in legends:
+            fig.add_layout(legend, 'right')
+        fig.legend.click_policy = 'hide'
+
+        return fig
 
 
 class LineSource:
